feat: add schema service and refactor trade system to in-memory DB

+++ src/main/java/com/example/find_my_edge/core/backtest/service/TradeFieldService.java
@@ -1,7 +1,5 @@
 package com.example.find_my_edge.core.backtest.service;
 
-import com.example.find_my_edge.common.enums.ResponseState;
-import com.example.find_my_edge.common.response.ApiResponse;
 import com.example.find_my_edge.core.backtest.dto.FieldDataRequest;
 import com.example.find_my_edge.core.backtest.dto.TradeRecordsResponse;
 import com.example.find_my_edge.core.backtest.entity.Trade;
@@ -9,16 +7,13 @@ import com.example.find_my_edge.core.backtest.entity.TradeField;
 import com.example.find_my_edge.core.backtest.mapper.TradeFieldMapper;
 import com.example.find_my_edge.core.backtest.mapper.TradeMapper;
 import com.example.find_my_edge.core.backtest.repository.TradeRepository;
-import com.example.find_my_edge.common.exceptions.TradeNotFound;
+import com.example.find_my_edge.common.exceptions.TradeNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 
@@ -30,7 +25,7 @@ public class TradeFieldService {
     private final TradeMapper tradeMapper;
     private final TradeFieldMapper fieldMapper;
 
-    public ResponseEntity<ApiResponse<Object>> addTradeRecords(List<FieldDataRequest> fields) {
+    public TradeRecordsResponse addTradeRecords(List<FieldDataRequest> fields) {
         Trade trade = new Trade();
 
         for (FieldDataRequest dto : fields) {
@@ -41,37 +36,26 @@ public class TradeFieldService {
 
         Trade saved = tradeRepository.save(trade);
 
-        return ResponseEntity.ok(ApiResponse.builder()
-                                            .state(ResponseState.SUCCESS)
-                                            .httpStatus(HttpStatus.OK.value())
-                                            .message("Trade Record added successfully")
-                                            .data(tradeMapper.toResponse(saved))
-                                            .meta(Map.of("empty", false, "count", 1))
-                                            .build());
+        return tradeMapper.toResponse(saved);
+
     }
 
 
-    public ResponseEntity<ApiResponse<Object>> getTradeRecords() {
+    public List<TradeRecordsResponse> getTradeRecords() {
         List<Trade> trades = tradeRepository.findAll();
 
-        List<TradeRecordsResponse> records = trades.stream()
-                                                   .map(tradeMapper::toResponse)
-                                                   .toList();
+        return trades.stream()
+                     .map(tradeMapper::toResponse)
+                     .toList();
 
-        return ResponseEntity.ok(ApiResponse.builder()
-                                            .state(ResponseState.SUCCESS)
-                                            .httpStatus(HttpStatus.OK.value())
-                                            .message("Trade Records found successfully")
-                                            .data(records)
-                                            .meta(Map.of("empty", false, "count", records.size()))
-                                            .build());
+
     }
 
     @Transactional
-    public ResponseEntity<ApiResponse<Object>> updateTradeRecords(Long tradeId, List<FieldDataRequest> fields) {
+    public TradeRecordsResponse updateTradeRecords(Long tradeId, List<FieldDataRequest> fields) {
         Optional<Trade> tradeOptional = tradeRepository.findById(tradeId);
 
-        Trade trade = tradeOptional.orElseThrow(() -> new TradeNotFound("Trade not found with id " + tradeId));
+        Trade trade = tradeOptional.orElseThrow(() -> new TradeNotFoundException("Trade not found with id " + tradeId));
 
         trade.getFields()
              .clear();
@@ -85,25 +69,16 @@ public class TradeFieldService {
 
         Trade saved = tradeRepository.save(trade);
 
-        return ResponseEntity.ok(ApiResponse.builder()
-                                            .state(ResponseState.SUCCESS)
-                                            .httpStatus(HttpStatus.OK.value())
-                                            .message("Trade Record updated successfully")
-                                            .data(tradeMapper.toResponse(saved))
-                                            .meta(Map.of("empty", false, "count", 1))
-                                            .build());
+        return tradeMapper.toResponse(saved);
     }
 
     @Transactional
-    public ResponseEntity<ApiResponse<Object>> deleteTradeRecord(Long tradeId) {
-        tradeRepository.deleteById(tradeId);
-        return ResponseEntity.ok(ApiResponse.builder()
-                                            .state(ResponseState.SUCCESS)
-                                            .httpStatus(HttpStatus.OK.value())
-                                            .message("Trade Record deleted successfully")
-                                            .data(null)
-                                            .meta(Map.of("empty", true, "count", 0))
-                                            .build());
-    }
+    public void deleteTradeRecord(Long tradeId) {
 
+        if (!tradeRepository.existsById(tradeId)) {
+            throw new TradeNotFoundException("Trade not found with id " + tradeId);
+        }
+
+        tradeRepository.deleteById(tradeId);
+    }
 }
